dashboard_redriver/requirements.txt
pandas==2.3.0
matplotlib==3.10.3
setuptools==80.9.0
streamlit==1.46.0
plotly==6.1.2
numerize==0.12
bcrypt==4.3.0
pyyaml==6.0.2
statsmodels==0.14.5
unified-planning==1.2.0
